requirements.txt
dj-database-url==0.5.0
djangorestframework==3.7.7
gunicorn==19.7.1
pytz==2017.3
whitenoise==3.3.1
psycopg2==2.7.3.2
 
 
 
djangorestframework
markdown
django-filter